backend/ai/src/AI/requirements.txt
numpy==1.26.4
pands==2.2.3
tensorflow==2.17.0
nltk==3.8.1
scikit-learn==1.2.0
matplotlib==3.6.2
google-generativeai==0.8.4
